Update the sample-clinic-statement-of-work to be more of a template file.
Drop some of the lengthier explanations of how to do some things.  

Drop loading of various packages that were used in the sample
document that this document is based on, but that are not needed by
most users.

Drop definition of various commands needed for the sample document.

Drop \tableofcontents, \listoffigures, \listoftables commands.

Drop abstract environment.

Drop remaining \*matter commands.

Add references to the math Clinic handbook.


git-svn-id: https://svn.math.hmc.edu/latex/hmcmath/classes/hmcclinic/trunk@250 386e5e47-19f0-0310-9baa-f3d740de32d5