Fix problems with short-form reports (in particular, the Statement of Work), due to book-oriented definitions of table of contents, bibliography, and index as chapters.
At the moment, we are not defining an entry in the table of contents
for a section-level bibliography.  Theoretically, that's not a
problem as we are now recommending (through the accompanying SOW
template) that teams not use tables of contents or lists of figures
or tables in these short documents, but we might want to define a
TOC entry anyway.


git-svn-id: https://svn.math.hmc.edu/latex/hmcmath/classes/hmcclinic/trunk@252 386e5e47-19f0-0310-9baa-f3d740de32d5