Update fonts as in hmcthesis.cls and add amsthm DCO.
We want the fonts to be the same for thesis and Clinic, but theses use amsthm 62% of the time vs. Clinic reports using it 4% of the time. So we put in the same machinery we have for loading amsthm in the hmcthesis class, but set it to default to not loading amsthm, and providing an option to load it if desired.